PR feedback, fix for XSS
